rename trame_setup.py to server.py
